Fix invalid use of glColorTableEXT on non-Windows platforms
We currently only define this and enable the paletted textures
functionality on Windows, so trying to use it elsewhere can lead to an
undefined symbol. I don't know why this code isn't enabled elsewhere,
but it should probably be handled by GLU. Having said that, I don't
think any real world code actually relies on this.